# pkg/api/rest/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Software migration management module",
        "title": "CM-Grasshopper REST API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "latest"
    },
    "basePath": "/grasshopper",
    "paths": {
        "/job/log/{jobId}": {
            "get": {
                "description": "Get k8s migration job log by job ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] K8s migration job APIs"
                ],
                "summary": "Get Job Log",
                "operationId": "get-job-log",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the job.",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully got the job log.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.JobLogRes"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the job log.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/job/status": {
            "get": {
                "description": "List k8s migration jobs.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] K8s migration job APIs"
                ],
                "summary": "List Job Status",
                "operationId": "list-job-status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number.",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Rows per page.",
                        "name": "row",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully listed jobs.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.JobExecution"
                            }
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to list jobs.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/job/status/{jobId}": {
            "get": {
                "description": "Get k8s migration job status by job ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "[Migration] K8s migration job APIs"
                ],
                "summary": "Get Job Status",
                "operationId": "get-job-status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the job.",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully got the job status.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.JobExecution"
                        }
                    },
                    "400": {
                        "description": "Sent bad request.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to get the job status.",
                        "schema": {
                            "$ref": "#/definitions/github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_cloud-barista_cm-grasshopper_pkg_api_rest_common.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.JobExecution": {
            "type": "object",
            "properties": {
                "error_message": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "job_type": {
                    "type": "string"
                },
                "log_path": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "metadata": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "resource_name": {
                    "type": "string"
                },
                "resource_type": {
                    "type": "string"
                },
                "source_cluster_name": {
                    "type": "string"
                },
                "source_namespace": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_cluster_name": {
                    "type": "string"
                },
                "target_namespace": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_cloud-barista_cm-grasshopper_pkg_api_rest_model.JobLogRes": {
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string"
                },
                "log": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}